requirements.txt
absl-py==2.2.2
ale-py==0.11.0
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
farama-notifications==0.0.4
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.0
grpcio==1.71.0
gymnasium==1.1.1
jinja2==3.1.6
kiwisolver==1.4.8
markdown==3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==6.31.0
psutil==7.0.0
pygame==2.6.1
pyglet==1.5.31
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
rich==14.0.0
six==1.17.0
stable-baselines3==2.6.0
stable-retro==0.9.5
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.7.0
tqdm==4.67.1
triton==3.3.0
typing-extensions==4.13.2
tzdata==2025.2
werkzeug==3.1.3